+++ package.json
@@ -125,12 +125,12 @@
     "fix-path": "2.1.0",
     "get-port": "3.2.0",
     "graphql": "0.13.2",
-    "graphql-code-generator": "0.14.2",
-    "graphql-codegen-typescript-apollo-angular": "0.14.2",
-    "graphql-codegen-typescript-common": "0.14.2",
-    "graphql-codegen-typescript-client": "0.14.2",
-    "graphql-codegen-typescript-resolvers": "0.14.2",
-    "graphql-codegen-typescript-server": "0.14.2",
+    "graphql-code-generator": "0.15.0-alpha.a7ec32b1",
+    "graphql-codegen-typescript-apollo-angular": "0.15.0-alpha.a7ec32b1",
+    "graphql-codegen-typescript-common": "0.15.0-alpha.a7ec32b1",
+    "graphql-codegen-typescript-client": "0.15.0-alpha.a7ec32b1",
+    "graphql-codegen-typescript-resolvers": "0.15.0-alpha.a7ec32b1",
+    "graphql-codegen-typescript-server": "0.15.0-alpha.a7ec32b1",
     "graphql-tag": "2.10.0",
     "jasmine-core": "3.3.0",
     "jasmine-marbles": "0.4.0",
